requirements.txt
huggingface-hub==0.19.4
transformers==4.35.2
sentence-transformers==2.2.2
gradio==4.8.0
chromadb==0.4.18
langchain==0.0.350
python-dotenv==1.0.0